requirements.txt
ruff
mypy
bandit
pip-audit
pytest-cov